Create RoomNetworking so the server can do messages
I totally forgor that I can't do any of the Mirror syncvars or messages on the actual server script, so instead it'll have its own component that does all of that for it.
