+++ Assets/Scripts/Network/RoomNetworking.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eed8e456542de2442b1ba13d09a35afe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
