+++ Assets/Scripts/Network/TankRoomManager.cs
@@ -19,18 +19,22 @@ public class TankRoomManager : NetworkRoomManager
 
     Chat chatroom; //For sending join and disconnect messages on
 
-    [SyncVar(hook = nameof(SetHorniTank))]
-    public GameObject HorniTank = null; //When our one and only tank gets spawned, we assign it here.
-
-    //Implement delegate as an event for encapsulation. This prevents subscribers from clearing the delegate or invoking it themselves
-    public event OnChangeHorniTank OnChangeHorniTankEvent;
-    public delegate void OnChangeHorniTank(GameObject HorniTank); //Delegate to fire when we change the tank reference
+    /// <summary>
+    /// A component on the Manager's gameobject that allows it to send messages back and forth. I dislike that the NetworkManager can't do that on it's own.
+    /// </summary>
+    public RoomNetworking RoomNetworking;
 
     public Dictionary<NetworkIdentity, PlayerProfile> connectedPlayers = new(); //Anybody added to the server gets stashed in here for me to use instead of Mirror's stuff. I know it's duplicated, but let me cook
 
     public Dictionary<ProfileGroup, GameObject> GroupToVehicles = new(); //This is for supporting multiple tanks later on
 
     public static new TankRoomManager singleton => NetworkManager.singleton as TankRoomManager;
+
+    public override void Awake()
+    {
+        base.Awake(); //I don't know what the manager does on awake, so just let it do its thing I guess
+        RoomNetworking = GetComponent<RoomNetworking>(); //Nab the component so we can send messages later
+    }
 
     public override void ReadyStatusChanged()
     {
@@ -99,10 +103,10 @@ public class TankRoomManager : NetworkRoomManager
         {
             //Time to spawn the tank in
             //We'll worry about picking a proper spawn point later on
-            //Debug.Log("Server moved into gameplay scene, spawning tank");
-            HorniTank = GameObject.Instantiate(horniTankPrefab); //Double check this puts the tank at 0,0,0
-            NetworkServer.Spawn(HorniTank);
-            OnChangeHorniTankEvent?.Invoke(HorniTank); //Only invoke the horniTank event if there are subscribers
+            
+            //Since the HorniTank is a syncvar we shouldn't need to manually invoke the event handler. This should just work
+            RoomNetworking.HorniTank = GameObject.Instantiate(horniTankPrefab); //Double check this puts the tank at 0,0,0
+            NetworkServer.Spawn(RoomNetworking.HorniTank); //Spawn the tank across the server for everyone
 
             GameObject canvas = GameObject.Find("Canvas");
             GameObject chatroomObject = GameObject.Instantiate(chatroomPrefab, canvas.transform);
@@ -134,12 +138,6 @@ public class TankRoomManager : NetworkRoomManager
         }
         //TODO: I want to add random connection messages like discord does with people joining a server. There is a trello card for this request.
         chatroom.SendServerMessage($"{profile.PlayerName} has connected!", Chat.MessageTypes.SERVER);
-    }
-
-    void SetHorniTank(GameObject oldTank, GameObject newTank)
-    {
-        Debug.Log("HorniTank has been updated, SyncVar hook called on client");
-        OnChangeHorniTankEvent?.Invoke(newTank);
     }
 
     public override void OnGUI()
